Adds a script for generating detections for Deep SORT through YOLO, tries to fix IoU calculation
The IoU score was wrong before because it was comparing the ground truth box to the YOLO resized prediction bounding box size. The bounding box can be scaled back up to original size but I am not certain that the algorithm that I implemented is correct. The IoU score however does appear to much more in line with what is expected.